back-end-app/main-app/requirements.txt
flask
flask_cors
flask_sqlalchemy
flask_restful
pika
psycopg2-binary
Werkzeug
tqdm